fix : error clean, refactoring - 1
