+++ YogiYoCloneIOS/YogiYoCloneIOS/ViewController/ViewControllers/HomeVC/Byun/DetailVC/Extension/UITableViewDelegate.swift
@@ -1,0 +1,28 @@
+//
+//  UITableViewDelegate.swift
+//  YogiYoCloneIOS
+//
+//  Created by Qussk_MAC on 2020/10/19.
+//  Copyright © 2020 김동현. All rights reserved.
+//
+
+import UIKit
+
+//MARK:-UITableViewDelegate
+extension DetailMenuVC: UITableViewDelegate {
+  
+  //헤더뷰 높이
+  func tableView(_ tableView: UITableView, heightForHeaderInSection section: Int) -> CGFloat {
+    section == 0 ? 0 : 1
+  }
+  
+  func tableView(_ tableView: UITableView, titleForHeaderInSection section: Int) -> String? {
+    section == 0 ? " " : " "
+  }
+  
+  func tableView(_ tableView: UITableView, heightForRowAt indexPath: IndexPath) -> CGFloat {
+    return UITableView.automaticDimension
+  }
+}
+
+
